requirements.txt
pillow
numpy
bs4
requests